requirements.txt
openai
beautifulsoup4
requests
tiktoken
pinecone-client>=2.2.4
pydantic
python-dotenv
